Branched Ciscowipe and tech program
Officially took the class for tech'ing stuff and moved it to it's own file. I have started to work on an argument parser to set up the tech tool as a cli tool that takes the asset list/ assets as an argument and then techs them in Makor depending on the parameters defined prior in the command.