Began file asset ingesting logic
Started to write the logic needed for ingesting a list of assets and looking for all of the flags that may be listed, and then taking corresponding action to make those changes in  Makor.  Need to rethink how I am doing it, because I don't want the logic to be in main() but it can't be above, may make it a class to be called on in main(). Need to figure out how to make multiple connections on multiple threads and get a different asset and attribute list to each thread. Also need to talk to Management about how many connections would be optimal.