Initial round of logic built
This commit mainly finished off most of the logic as far as I know for the actual execution of teching assets and putting them in Makor, it added the threading, although I do need to figure out switching Tech_stuff into a callable function, I will probably just call the object in another function and pipe the args around and stuff, itll be inneficient, but it'll work for now. I have tidied up the logic to split each line from any potential files and splitting the arguments from the actual asset, so the main part of the program can actually do the basic teaching of the asset, and then it calls a separate function to process any arguments, if there are any. I have split it up so for now there will be 4 connections to the server at any 1 time, ill see if the server can handle more so itll be sped up more with more connections at a time and thus more assets being processed simultaneously.