Started figuring out how to do PyQt
Decided that if I am going to be doing multiple serial instances at once, then I am going to need a GUI, its going to be the easiest to coordinate multiple sessions. So this commit is me messing around with some classes , moving the delete functions around, and barely starting the GUI application code in main().